requirements.txt
tqdm
unstructured
unstructured[pdf]
langchain-community~=0.3.0
langchain~=0.3.0
sentence-transformers
faiss-cpu
torch==2.0.1+cu117 torchvision==0.20.1+cu117 -f https://download.pytorch.org/whl/torch_stable.html
torchaudio==2.5.1
unsloth~=2024.8
datasets~=3.0.0
bitsandbytes-windows
transformers~=4.44.2
pdfplumber~=0.11.4
sentencepiece
accelerate
protobuf
huggingface-hub~=0.26.2
pandas~=2.2.3
PyPDF2~=3.0.1
